chore: check for warnings and resolve existing ones

+++ src/stream/fileSegment.cpp
@@ -46,7 +46,7 @@ constexpr int32_t MAGIC_AND_VERSION = MAGIC_BYTES << 8 | VERSION;
 
 #pragma pack(push, 4)
 struct LogEntryHeader {
-    int32_t magic_and_version = static_cast<int32_t>(_htonl(MAGIC_AND_VERSION));
+    int32_t magic_and_version = static_cast<int32_t>(_htonl(static_cast<std::uint32_t>(MAGIC_AND_VERSION)));
     int32_t relative_sequence_number = 0;
     int32_t byte_position = 0;
     int64_t crc = 0;
@@ -143,11 +143,12 @@ StreamError FileSegment::open(bool full_corruption_check_on_open) {
             continue;
         }
         if (full_corruption_check_on_open) {
-            auto value_or = read(header.relative_sequence_number + _base_seq_num, ReadOptions{
-                                                                                      .check_for_corruption = true,
-                                                                                      .may_return_later_records = false,
-                                                                                      .suggested_start = offset,
-                                                                                  });
+            auto value_or = read(static_cast<std::uint64_t>(header.relative_sequence_number) + _base_seq_num,
+                                 ReadOptions{
+                                     .check_for_corruption = true,
+                                     .may_return_later_records = false,
+                                     .suggested_start = offset,
+                                 });
             if (!value_or) {
                 truncateAndLog(offset, value_or.err());
                 continue;
@@ -155,9 +156,10 @@ StreamError FileSegment::open(bool full_corruption_check_on_open) {
         }
 
         offset += HEADER_SIZE;
-        offset += header.payload_length_bytes;
-        _total_bytes += header.payload_length_bytes + HEADER_SIZE;
-        _highest_seq_num = std::max(_highest_seq_num, _base_seq_num + header.relative_sequence_number);
+        offset += static_cast<uint32_t>(header.payload_length_bytes);
+        _total_bytes += static_cast<std::uint32_t>(header.payload_length_bytes) + HEADER_SIZE;
+        _highest_seq_num =
+            std::max(_highest_seq_num, _base_seq_num + static_cast<std::uint64_t>(header.relative_sequence_number));
     }
 }
 
@@ -166,25 +168,27 @@ LogEntryHeader FileSegment::convertSliceToHeader(const OwnedSlice &data) {
     // Use memcpy instead of reinterpret cast to avoid UB.
     memcpy(&header, data.data(), sizeof(LogEntryHeader));
 
-    header.payload_length_bytes = static_cast<int32_t>(_ntohl(header.payload_length_bytes));
-    header.relative_sequence_number = static_cast<int32_t>(_ntohl(header.relative_sequence_number));
-    header.byte_position = static_cast<int32_t>(_ntohl(header.byte_position));
-    header.crc = static_cast<int64_t>(_ntohll(header.crc));
-    header.timestamp = static_cast<int64_t>(_ntohll(header.timestamp));
-    header.magic_and_version = static_cast<int32_t>(_ntohl(header.magic_and_version));
+    header.payload_length_bytes = static_cast<int32_t>(_ntohl(static_cast<std::uint32_t>(header.payload_length_bytes)));
+    header.relative_sequence_number =
+        static_cast<int32_t>(_ntohl(static_cast<std::uint32_t>(header.relative_sequence_number)));
+    header.byte_position = static_cast<int32_t>(_ntohl(static_cast<std::uint32_t>(header.byte_position)));
+    header.crc = static_cast<int64_t>(_ntohll(static_cast<std::uint64_t>(header.crc)));
+    header.timestamp = static_cast<int64_t>(_ntohll(static_cast<std::uint64_t>(header.timestamp)));
+    header.magic_and_version = static_cast<int32_t>(_ntohl(static_cast<std::uint32_t>(header.magic_and_version)));
     return header;
 }
 
 expected<uint64_t, FileError> FileSegment::append(BorrowedSlice d, int64_t timestamp_ms, uint64_t sequence_number,
                                                   bool sync) {
-    auto ts = static_cast<int64_t>(_htonll(timestamp_ms));
+    auto ts = static_cast<int64_t>(_htonll(static_cast<std::uint64_t>(timestamp_ms)));
     auto data_len_swap = static_cast<int32_t>(_htonl(d.size()));
     auto byte_position = static_cast<int32_t>(_htonl(_total_bytes));
 
     auto crc = static_cast<int64_t>(_htonll(
         crc32::crc32_of(BorrowedSlice{&ts, sizeof(ts)}, BorrowedSlice{&data_len_swap, sizeof(data_len_swap)}, d)));
     auto header = LogEntryHeader{
-        .relative_sequence_number = static_cast<int32_t>(_htonl(static_cast<int32_t>(sequence_number - _base_seq_num))),
+        .relative_sequence_number =
+            static_cast<int32_t>(_htonl(static_cast<std::uint32_t>(sequence_number - _base_seq_num))),
         .byte_position = byte_position,
         .crc = crc,
         .timestamp = ts,
@@ -257,13 +261,14 @@ expected<OwnedRecord, StreamError> FileSegment::read(uint64_t sequence_number, c
         // We found the one we want, or the next available sequence number was acceptable to us
         if (header.relative_sequence_number == expected_rel_seq_num ||
             (header.relative_sequence_number > expected_rel_seq_num && read_options.may_return_later_records)) {
-            auto data_or = _f->read(offset + HEADER_SIZE, offset + HEADER_SIZE + header.payload_length_bytes);
+            auto data_or = _f->read(offset + HEADER_SIZE,
+                                    offset + HEADER_SIZE + static_cast<std::uint32_t>(header.payload_length_bytes));
             if (!data_or) {
                 return StreamError{StreamErrorCode::ReadError, header_data_or.err().msg};
             }
             auto data = std::move(data_or.val());
-            auto data_len_swap = static_cast<int32_t>(_htonl(header.payload_length_bytes));
-            auto ts_swap = static_cast<int64_t>(_htonll(header.timestamp));
+            auto data_len_swap = static_cast<int32_t>(_htonl(static_cast<std::uint32_t>(header.payload_length_bytes)));
+            auto ts_swap = static_cast<int64_t>(_htonll(static_cast<std::uint64_t>(header.timestamp)));
 
             if (read_options.check_for_corruption &&
                 header.crc != static_cast<int64_t>(crc32::crc32_of(BorrowedSlice{&ts_swap, sizeof(ts_swap)},
@@ -281,7 +286,7 @@ expected<OwnedRecord, StreamError> FileSegment::read(uint64_t sequence_number, c
         }
 
         offset += HEADER_SIZE;
-        offset += header.payload_length_bytes;
+        offset += static_cast<std::uint32_t>(header.payload_length_bytes);
     }
 }
